Added option to return a string from restore
I usually don't care about how many calls were made to `console.log`, so returning an array from `restore` isn't helpful. Of course I can add `.join('')` to the result, but it is much cleaner if there is an added boolean parameter to `restore`, which if truthy, returns a string.